pyproject.toml
[project]
name = "textbook-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.26.0",
    "bitsandbytes>=0.45.5",
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.24",
    "manim>=0.19.0",
    "numpy>=2.2.5",
    "ollama>=0.4.8",
    "pydantic>=2.11.3",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.44.1",
    "transformers>=4.51.3",
]
